requirements.txt
django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
python-dotenv==1.0.1
django-celery-beat==2.5.0 
gunicorn==20.1.0 
